trajets etage 1 et correction bugs

+++ app/src/main/java/com/example/visiteguideecegep/Intersection.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 
 public class Intersection {
     int numero;
-    double x;
-    double y;
+    float x;
+    float y;
     boolean utilise;
     boolean escalier;
     ArrayList<Integer> numerosIntersectionsReliees = new ArrayList<>();
